If a group key is provided try to find file using that key
The default behavior is to try to match selected files by looking for the
detector name in the file name. If the "group" key is present on the detector
we will now also check for the presence of that value in the file name in an
attempt to automatically select files and match them to their detector.

Signed-off-by: Brianna Major <brianna.major@kitware.com>