Collapse and hide non-matching xrs
If an instrument has multiple x-ray sources then when we are in polar mode we
will automatically hide any masks created for a different x-ray source than the
current selection. Those sections are also automatically collapsed in the
MaskManagerDialog and the section with masks that match the current x-ray
source is expanded.

Signed-off-by: Brianna Major <brianna.major@kitware.com>